refactor(core): Move base event classes to eventbus
- Update subclasses and references
- Remove "send" method from BaseEvent that returns a Promise
