+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/events/RepositoryEvent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2016 B2i Healthcare Pte Ltd, http://b2i.sg
+ * Copyright 2011-2022 B2i Healthcare Pte Ltd, http://b2i.sg
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,6 +16,8 @@
 package com.b2international.snowowl.core.events;
 
 import static com.google.common.base.Preconditions.checkNotNull;
+
+import com.b2international.snowowl.eventbus.events.SystemNotification;
 
 /**
  * @since 4.5
