+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/jobs/RemoteJobNotification.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 B2i Healthcare Pte Ltd, http://b2i.sg
+ * Copyright 2017-2022 B2i Healthcare Pte Ltd, http://b2i.sg
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,7 +19,7 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.Set;
 
-import com.b2international.snowowl.core.events.SystemNotification;
+import com.b2international.snowowl.eventbus.events.SystemNotification;
 
 /**
  * @since 5.7
