+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/messaging/MessageNotification.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 B2i Healthcare Pte Ltd, http://b2i.sg
+ * Copyright 2018-2022 B2i Healthcare Pte Ltd, http://b2i.sg
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,7 @@
  */
 package com.b2international.snowowl.core.messaging;
 
-import com.b2international.snowowl.core.events.SystemNotification;
+import com.b2international.snowowl.eventbus.events.SystemNotification;
 
 /**
  * @since 7.0
