+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/validation/ValidationDeleteNotification.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2020 B2i Healthcare Pte Ltd, http://b2i.sg
+ * Copyright 2019-2022 B2i Healthcare Pte Ltd, http://b2i.sg
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,7 +17,7 @@ package com.b2international.snowowl.core.validation;
 
 import java.util.Set;
 
-import com.b2international.snowowl.core.events.SystemNotification;
+import com.b2international.snowowl.eventbus.events.SystemNotification;
 
 /**
  * @since 6.20.0
