+++ netty/com.b2international.snowowl.eventbus/META-INF/MANIFEST.MF
@@ -12,6 +12,7 @@ Require-Bundle: com.google.guava;bundle-version="[27.1.0,28.0.0)",
  io.netty.transport;visibility:=reexport,
  io.netty.buffer;visibility:=reexport
 Export-Package: com.b2international.snowowl.eventbus,
+ com.b2international.snowowl.eventbus.events,
  com.b2international.snowowl.eventbus.netty,
  com.b2international.snowowl.internal.eventbus;x-internal:=true,
  com.b2international.snowowl.internal.eventbus.netty;x-internal:=true
