+++ netty/com.b2international.snowowl.eventbus/src/com/b2international/snowowl/eventbus/events/Event.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2015 B2i Healthcare Pte Ltd, http://b2i.sg
+ * Copyright 2011-2022 B2i Healthcare Pte Ltd, http://b2i.sg
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.b2international.snowowl.core.events;
+package com.b2international.snowowl.eventbus.events;
 
 import java.io.Serializable;
 
-import com.b2international.snowowl.core.events.util.Promise;
 import com.b2international.snowowl.eventbus.IEventBus;
 import com.b2international.snowowl.eventbus.IHandler;
 import com.b2international.snowowl.eventbus.IMessage;
@@ -34,7 +33,5 @@ public interface Event extends Serializable {
 	void publish(IEventBus bus);
 	
 	void send(IEventBus bus, IHandler<IMessage> replyHandler);
-	
-	<T> Promise<T> send(IEventBus bus, Class<T> returnType);
 	
 }
